+++ packages/types/src/api/params.ts
@@ -60,6 +60,7 @@ export enum ApiParams {
 	headHash = "headHash",
 	behavioralData = "behavioralData",
 	simdReadings = "simdReadings",
+	currentUrl = "currentUrl",
 	decisionMachineScope = "decisionMachineScope",
 	decisionMachineRuntime = "decisionMachineRuntime",
 	decisionMachineSource = "decisionMachineSource",