+++ integration/frameworks/vue/vue-procaptcha-wrapper/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @prosopo/vue-procaptcha-wrapper
 
+## 1.1.90
+### Patch Changes
+
+  - @prosopo/procaptcha-wrapper@2.6.92
+
 ## 1.1.89
 ### Patch Changes