+++ packages/redis-client/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @prosopo/redis-client
 
+## 1.0.27
+### Patch Changes
+
+  - @prosopo/logger@2.0.1
+
 ## 1.0.26
 ### Patch Changes